backend/training-service/requirements.txt
kafka-python
psutil
numpy
tensorflow
